Separate createStore
