+++ examples/redux/gatsby-browser.js
@@ -1,23 +1,11 @@
 import React from 'react'
 import { BrowserRouter as Router } from 'react-router-dom'
-
 import { Provider } from 'react-redux'
-import { createStore } from 'redux'
 
-const reducer = (state, action) => {
-    console.log(action)
-    if (action.type === `INCREMENT`) {
-        return Object.assign({}, state, {
-            count: state.count + 1,
-        })
-    }
-    return state
-}
-
-const initialState = { count: 0 }
+import createStore from './src/state/createStore'
 
 exports.replaceRouterComponent = ({ history }) => {
-    const store = createStore(reducer, initialState)
+    const store = createStore()
 
     const ConnectedRouterWrapper = ({ children }) => (
         <Provider store={store}>
